Edit Product model

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/store/dto/ProductDto.java
@@ -7,9 +7,10 @@ import ru.yandex.practicum.store.enums.ProductCategory;
 import ru.yandex.practicum.store.enums.ProductState;
 import ru.yandex.practicum.store.enums.QuantityState;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record ProductDto(UUID productId, @NotBlank String productName, @NotBlank String description,
                          @NotBlank String imageSrc, @NotNull QuantityState quantityState,
                          @NotNull ProductState productState, @NotNull ProductCategory productCategory,
-                         @NotNull @Positive Float price) {}
+                         @NotNull @Positive BigDecimal price) {}
