+++ commerce/shopping-store/src/main/java/ru/yandex/practicum/model/Product.java
@@ -7,6 +7,7 @@ import ru.yandex.practicum.store.enums.ProductCategory;
 import ru.yandex.practicum.store.enums.ProductState;
 import ru.yandex.practicum.store.enums.QuantityState;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Getter
@@ -43,5 +44,5 @@ public class Product {
     ProductCategory productCategory;
 
     @Column(name = "price")
-    Float price;
+    BigDecimal price;
 }
